Keep runtime tests beside their owners
Remove duplicate validation paths and relocate focused file-command, needs, secrets, and workflow tests out of the runtime catch-all.

Part of PB-3178.

Amp-Thread-ID: https://ampcode.com/threads/T-01a060ce-cc2b-737a-8463-595c3e3587cb
